requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
Pillow
firebase-admin
resend
requests
mistune
mediapipe
opencv-python
jinja2
psycopg[binary]
